Return undef from read_body
The return values of read_(content|chunked)_body are
different and in both cases, they are ignored by
code that calls read_body.  If read_body is to
return something, it should be consistent regardless of
the underlying read.  Since nothing uses the return
currently, this code change makes that obvious.