Add XXX comments for trailing headers
In reading, trailing headers are discarded.  Is
this OK or should the API/code change to make
use of trailing headers.  (Do they merge with
or overwrite original headers?)

In writing, there is no code that provides trailing
headers.  Should the feature be removed?  Or
should the API change to allow trailing headers
via some sort of callback?