Possible monkeypatch for testing
The idea is to monkeypatch HTTP::Tiny::Handle to mock the server
iteraction.  First, two temporary filehandles would be opened.  One is
blank to receive the request and the other one has the response.  The
connect() function is mocked to set up the blank handle as the
filehandle that is written to.  Then write_request() is wrapped
to swap in the response filehandle after the original write_request
completes.  Additionally, both can_* functions are mocked to return
true as select() on filehandles doesn't work on Windows.

Rough (untested) example:

    use HTTP::Tiny;
    use t::Util;
    my $req_fh = tmpfile
    my $res_fh = tmpfile( $response_text );
    set_socket_source( $req_fh, $res_fh );
    monkey_patch();

    # proceed with testing. Call set_socket_source as needed with fresh
    # handles