Specify hostname for SNI, and permit setting the SSL_verifycn_name
This permits users to set SSL_verifycn_name, in case they want to
mess with validation of the SSL certificate's CN against the domain
we're talking to. Otherwise, we verify the CN against the domain
name we expected to be talking to, which is recommended.

We also set SSL_hostname for SNI support unless the user already did.

Addresses some of dagolden/p5-http-tiny#5