Added support for persistent connections
If keep_alive => 1, then HTTP::Tiny will try to reuse the previous socket,
but only for the same scheme/host/port/timeout. Any other changes will
use a new socket

It won't reuse the connection if the content-length header is missing or
incorrect.

If r/w attributes are modified, the cached handle will also be cleared.