Ensure tests pass with a SOCKS proxy configured
HTTP::Tiny uses "ALL_PROXY" and "all_proxy" from the environment to
configure which proxy it uses.  This change makes tests pass on a system
with these values set.

HTTP::Tiny throws an exception if it has a value of 0 or the empty
string for these environment variables, so always set them to the
undefined value.