Add develop dependency for non-core Perl::Critic policy
Lax::ProhibitStringyEval::ExceptForRequire is listed in perlcritic.rc
but it isn't a core Perl::Critic policy. If it is not installed, a
warning is triggered by the Test::Perl::Critic author test.
DZP::Test::Perl::Critic doesn't yet parse perlcritic.rc to inject the
missing policiy (I plan to fix that), so for now, just add the
dependency manually.