SSL environment controls, and warn about insecure connections
The verify_SSL option is disabled by default, which is both a
vulnerability and an RFC 2818 violation.  Being a heavily-used
module by many in-house scripts, it is probably not practical
to change this default, but users need to be warned so that they
can apply whatever security intentions they have rather than
blindly receiving an insecure default.

This patch changes the verify_SSL attribute to remain 'undef'
if the user does not take any steps to assign a value to it.
This condition can then detect cases where an SSL connection
is being established without verification and where the user
did not op-in to this behavior.  It generates one warning
using 'warn', and then suppresses future warnings.

The first new environment variable HTTP_TINY_VERIFY_SSL_DEFAULT
is a boolean that can either opt-in to the default of false, or
change the default to true.  Both settings suppress the warning.

The second environment variable HTTP_TINY_VERIFY_SSL is a
boolean that can override the attribute chosen by users of the
module.  This will become necessary if module authors who use
HTTP::Tiny begin including the (verify_SSL => 1) setting in the
construction of the user agent but end-users of those modules
want to override the decision to allow insecure connections.
This mirrors the utility of PERL_LWP_SSL_VERIFY_HOSTNAME which
end-users can set to quickly permit insecure connections with
LWP::UserAgent.